+++ app/Http/Controllers/IndustryExperienceController.php
@@ -10,59 +10,70 @@ use Illuminate\Support\Facades\Storage;
 
 class IndustryExperienceController extends Controller
 {
-
     private $fileExtension = '.pdf';
+    private $methodMappingTable = array(
+        'index' => 'viewAny',
+        'show' => 'view',
+        'create' => 'create',
+        'store' => 'create',
+        'edit' => 'update',
+        'update' => 'update',
+        'destroy' => 'delete',
+    );
 
     public function index()
     {
+        $this->authorize($this->methodMappingTable['index'], Industry_experience::class);
         $collection = Industry_experience::where('username', Auth::user()->username)->get();
         return view('industry_experience.index', compact('collection'));
     }
 
     public function create()
     {
+        $this->authorize($this->methodMappingTable['create'], Industry_experience::class);
         return view('industry_experience.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize($this->methodMappingTable['store'], Industry_experience::class);
         $data = $this->validation($request);
         $data['username'] = Auth::user()->username;
         Industry_experience::create($data);
         return redirect()->route('industry_experience.index');
     }
 
-    public function destroy($id)
+    public function destroy(Industry_experience $industry_experience)
     {
-        $row = Industry_experience::where('id', $id)->firstOrFail();
-        $oldIdentification = $row->identification;
+        $this->authorize($this->methodMappingTable['destroy'], $industry_experience);
+        $oldIdentification = $industry_experience->identification;
         Storage::delete('public/industry_experience/' . $oldIdentification . $this->fileExtension);
-        $row->delete();
+        $industry_experience->delete();
         return redirect()->route('industry_experience.index');
     }
 
-    public function edit($id)
+    public function edit(Industry_experience $industry_experience)
     {
-        $collection = Industry_experience::where('id', $id)->firstOrFail();
-        return view('industry_experience.edit', compact('collection'));
+        $this->authorize($this->methodMappingTable['edit'], $industry_experience);
+        return view('industry_experience.edit', compact('industry_experience'));
     }
 
-    public function show($id)
+    public function show(Industry_experience $industry_experience)
     {
-        $collection = Industry_experience::where('id', $id)->firstOrFail();
-        return view('industry_experience.show', compact('collection'));
+        $this->authorize($this->methodMappingTable['show'], $industry_experience);
+        return view('industry_experience.show', compact('industry_experience'));
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, Industry_experience $industry_experience)
     {
+        $this->authorize($this->methodMappingTable['update'], $industry_experience);
         $data = $this->validation($request);
         $data['username'] = Auth::user()->username;
-        $row = Industry_experience::where('id', $id);
         if (isset($data['identification'])) {
-            $oldIdentification = $row->identification;
+            $oldIdentification = $industry_experience->identification;
             Storage::delete('public/industry_experience/' . $oldIdentification . $this->fileExtension);
         }
-        $row->update($data);
+        $industry_experience->update($data);
         return redirect()->route('industry_experience.index');
     }
